管理员端更改数据库连接
已由superbase改为postgre sql，并附上postgre sql初始化的.md文档
双击启动管理员端.bat即可使用管理员端

+++ backend/src/main/resources/schema.sql
@@ -270,7 +270,7 @@ create table if not exists reviews (
     task_id uuid not null,
     from_user_id uuid not null,
     to_user_id uuid not null,
-    rating smallint not null,
+    rating integer not null,
     comment varchar(500),
     created_at timestamp with time zone not null default now(),
     updated_at timestamp with time zone not null default now(),
